Rename files from projectApp to workspaceApp
