+++ packages/server/src/sdk/index.ts
@@ -15,7 +15,7 @@ import * as screens from "./app/screens"
 import * as common from "./app/common"
 import * as oauth2 from "./app/oauth2"
 import * as ai from "./app/ai"
-import * as projectApps from "./app/projectApps"
+import * as projectApps from "./app/workspaceApps"
 
 const sdk = {
   backups,
