+++ packages/types/src/api/web/app/index.ts
@@ -11,7 +11,7 @@ export * from "./layout"
 export * from "./metadata"
 export * from "./oauth2"
 export * from "./permission"
-export * from "./projectApp"
+export * from "./workspaceApp"
 export * from "./query"
 export * from "./role"
 export * from "./rowAction"
