+++ packages/types/src/documents/app/index.ts
@@ -9,7 +9,7 @@ export * from "./layout"
 export * from "./links"
 export * from "./metadata"
 export * from "./oauth2"
-export * from "./projectApp"
+export * from "./workspaceApp"
 export * from "./query"
 export * from "./role"
 export * from "./row"
